+++ docs-app/app/docs-support.gts
@@ -1,7 +1,9 @@
+import { highlight } from 'docs-app/components/highlight';
+import { Compiled, defaultOptions } from 'docs-app/markdown';
 import { RemoteData } from 'ember-resources/util/remote-data';
 
 import type { TOC } from '@ember/component/template-only';
-import type { DeclarationReflection,ReferenceType,ReflectionGroup, SomeType } from 'typedoc';
+import type { DeclarationReference,DeclarationReflection,ReferenceType,ReflectionGroup, SomeType } from 'typedoc';
 
 const infoFor = (data: ReflectionGroup, module: string, name: string) => {
   let found = data.children
@@ -11,6 +13,13 @@ const infoFor = (data: ReflectionGroup, module: string, name: string) => {
   return found as DeclarationReflection | undefined;
 };
 
+/**
+  * Assumptions:
+  * - we are documenting public API
+  *   - component properties and methods are not public API
+  *     - including the constructor, inherited methods, etc
+  *   - only the signature describes what the public API is.
+  */
 export const APIDocs: TOC<{
   Args: {
     module: string;
@@ -24,9 +33,39 @@ export const APIDocs: TOC<{
 
     {{#if request.value}}
       <Styles />
-      <Declaration @info={{infoFor request.value @module @name}} />
+
+      <section {{highlight request.value}}>
+        {{#let (infoFor request.value @module @name) as |info|}}
+          {{#if (isGlimmerComponent info)}}
+            <GlimmerComponent @info={{info}} />
+          {{else}}
+            <Declaration @info={{info}} />
+          {{/if}}
+
+        {{/let}}
+      </section>
     {{/if}}
   {{/let}}
+</template>;
+
+function isGlimmerComponent(info: DeclarationReference) {
+  let extended = info?.extendedTypes?.[0]
+
+  if (!extended) return false;
+
+  return extended.name === 'default' && extended.package === '@glimmer/component';
+}
+
+function typeArg(info: DeclarationReference) {
+  let extended = info?.extendedTypes?.[0]
+
+  if (!extended) return false;
+
+  return extended.typeArguments[0]
+}
+
+const GlimmerComponent: TOC<{ Args: { info: SomeType }}> = <template>
+  <Type @info={{typeArg @info}} />
 </template>;
 
 const join = (lines: string[]) => lines.join('\n');
@@ -35,19 +74,28 @@ const isIgnored = (name: string) => ['__type', 'TOC', 'TemplateOnlyComponent'].i
 const isConst = (x) => x.flags.isConst;
 const not = x => !x;
 
+const RenderedComment: TOC<{}> = <template>
+  {{#let (Compiled (join (text @text)) defaultOptions) as |compiled|}}
+    {{#if compiled.isReady}}
+      <compiled.component />
+    {{/if}}
+  {{/let}}
+</template>;
+
 const Declaration: TOC<{
   Args: {
     info: DeclarationReflection | undefined
   }
 }> = <template>
   {{#if @info}}
+    {{log @info.name @info}}
     {{#if (not (isIgnored @info.name))}}
       <h3 class="typedoc-declaration-name">{{@info.name}}</h3>
     {{/if}}
 
     {{#if (isConst @info)}}
       {{#if @info.comment.summary}}
-        {{join (text @info.comment.summary)}}
+        <RenderedComment @text={{@info.comment.summary}} />
       {{/if}}
     {{/if}}
 
@@ -60,10 +108,9 @@ const Declaration: TOC<{
         <li><Declaration @info={{child}} /></li>
       {{/each}}
     </ul>
-
     {{#if (not (isConst @info))}}
       {{#if @info.comment.summary}}
-        {{join (text @info.comment.summary)}}
+        <RenderedComment @text={{@info.comment.summary}} />
       {{/if}}
     {{/if}}
 
@@ -77,8 +124,10 @@ const Reflection: TOC<{ info: { declaration: DeclarationReflection }}> = <templa
 const isReference = (x: { type: string }) => x.type === 'reference';
 const isReflection = (x: { type: string }) => x.type === 'reflection';
 const isIntrinsic = (x: { type: string }) => x.type === 'intrinsic';
+const isTuple = (x: { type: string }) => x.type === 'tuple';
 
 const Reference: TOC<{ info: ReferenceType }> = <template>
+  {{log @info}}
   {{#if (not (isIgnored @info.name))}}
     <pre class="typedoc-reference">{{@info.name}}</pre>
   {{/if}}
@@ -91,6 +140,12 @@ const Intrinsic: TOC<{}> = <template>
   <span class="typedoc-intrinsic">{{@info.name}}</span>
 </template>;
 
+const Tuple: TOC<{ Args: { info: SomeType }}> = <template>
+  {{#each @info.elements as |element|}}
+    <Type @info={{element}} />
+  {{/each}}
+</template>;
+
 const Type: TOC<{ Args: { info: SomeType }}> = <template>
   {{#if (isReference @info)}}
     <Reference @info={{@info}} />
@@ -98,6 +153,8 @@ const Type: TOC<{ Args: { info: SomeType }}> = <template>
     <Reflection @info={{@info}} />
   {{else if (isIntrinsic @info)}}
     <Intrinsic @info={{@info}} />
+  {{else if (isTuple @info)}}
+    <Tuple @info={{@info}} />
   {{else}}
     {{log 'x' @info}}
   {{/if}}
