+++ ember-primitives/tsconfig.typedoc.json
@@ -4,7 +4,10 @@
     "baseUrl": "./"
   },
   "typedocOptions": {
-    "entryPoints": ["dist/*.d.ts"],
+    "entryPoints": [
+      "dist/*.d.ts",
+      "dist/components/**.d.ts"
+      ],
   }
   "include": ["dist/*.d.ts", "dist/**/*.d.ts"]
 }
